Initail Commit
Script was getting big enough that I felt the need to start tracking the coding process. Once things were slighly functional I knew I was gonna need this project in a Git repo.

While this particular script isn't quite at that point per-say, I wanted to make an inital commit now becuase I noticed this script was teaching me a lot; at a point I realized the functional specific content of this could be stripped and this could be turned into a "Bash BoilerPlate" script.

A "Bash BoilerPlate" could be implemented at LW to provide unifrom scirpt creation to provide a stepping stone for and to help foster novice scripters.